Customizing Bootstrap's Sass Settings
In your code editor, open up the file: /node_modules/bootstrap/scss/_custom.scss.
Notice how each of these sass variables has a !default at the end? That's specific to Sass and works opposite of the !important rule. If the variable has not been defined elsewhere, such as in the _custom sass file, then it will default to this value.